Labels for the tokens
