+++ apps/block_scout_web/assets/css/components/_tile.scss
@@ -578,3 +578,30 @@ $cube-quantity: 5;
 .dark-theme-applied .dark-block-loader {
   background-color: #313355;
 }
+
+.title-with-label {
+  display: inline-block;
+  vertical-align: middle;
+  margin-top: 3px;
+  @include media-breakpoint-down(sm) {
+    display: inline;
+  }
+}
+
+.view-original-token-container {
+  margin-top: 15px;
+  font-size: 12px;
+}
+
+.token-address-mobile {
+  @include media-breakpoint-down(sm) {
+    display: inline-block;
+    word-break: break-all;
+  }
+}
+
+.bridged-tokens-buttons-mobile {
+  @include media-breakpoint-down(sm) {
+    margin-top: 0.5em;
+  }
+}
