+++ apps/block_scout_web/lib/block_scout_web/templates/bridged_tokens/_tile.html.eex
@@ -34,7 +34,8 @@
         <div class="bs-label right <%= @bridged_token.type %>"><%= String.upcase(@bridged_token.type) %></div>
       <% end %>
       <%= if @bridged_token && @bridged_token.foreign_chain_id do %>
-        <div class="bs-label right ml-2 destination-<%= chain_id_display_name(@bridged_token.foreign_chain_id) %>"><%= String.upcase(chain_id_display_name(@bridged_token.foreign_chain_id)) %></div>
+        <% tag = Chain.chain_id_display_name(@bridged_token.foreign_chain_id) %>
+        <div class="bs-label right ml-2 destination-<%= tag %>"><%= String.upcase(tag) %></div>
       <% end %>
     </div>
   </td>
