+++ apps/block_scout_web/lib/block_scout_web/templates/bridged_tokens/index.html.eex
@@ -2,7 +2,8 @@
   <div class="card">
     <div class="card-body" data-async-load data-async-listing="<%= @current_path %>">
 	  <div class="tokens-list-search-input-outer-container">
-		<h1 class="card-title"><%= gettext "Bridged Tokens from " %><%= @chain %><div class="bs-label ml-2 destination-<%= chain_id_display_name(@chain_id) %>"><%= String.upcase(chain_id_display_name(@chain_id)) %></div></h1>
+		<% tag = Chain.chain_id_display_name(@chain_id) %>
+		<h1 class="card-title"><div class="title-with-label"><%= gettext "Bridged Tokens from " %><%= @chain %></div><div class="bs-label ml-2 destination-<%= tag %>"><%= String.upcase(tag) %></div></h1>
 		<label class="tokens-list-search-input-container">
 			<input data-search-field="" class="form-control tokens-list-search-input" type="text" name="filter" placeholder="Token name or symbol" id="search-text-input" />
 	  	</label>
