+++ apps/block_scout_web/lib/block_scout_web/templates/form/_tag.html.eex
@@ -1,0 +1,3 @@
+<div class="bs-label <%= @additional_classes |> Enum.join(" ") %>">
+    <%= @text %>
+</div>
