+++ apps/block_scout_web/lib/block_scout_web/templates/tokens/overview/_details.html.eex
@@ -20,13 +20,19 @@
               <% true -> %>
                 <%= nil %>
             <% end %>
+            <% foreign_chain_id = if Map.has_key?(@token, :foreign_chain_id), do: @token.foreign_chain_id, else: nil %>
+            <% tag = Chain.chain_id_display_name(foreign_chain_id) %>
             <%= if token_name?(@token) do %>
-              <%= @token.name %>
+                <div class="title-with-label"><%= @token.name %></div>
+                <%= if tag !== "" do %>
+                  <%= render BlockScoutWeb.FormView, "_tag.html", text: "bridged", additional_classes: ["bridged", "ml-1"] %>
+                  <%= render BlockScoutWeb.FormView, "_tag.html", text: String.upcase(tag), additional_classes: ["destination-#{tag}", "ml-1"] %>
+                <% end %>
             <% else %>
               <%= gettext("Token Details") %>
             <% end %>
             <!-- buttons -->
-            <span class="overview-title-buttons token float-right">
+            <span class="overview-title-buttons token float-right <%= if @token.bridged, do: "bridged-tokens-buttons-mobile", else: "" %>">
               <span class="overview-title-item" data-clipboard-text="<%= Address.checksum(@token.contract_address_hash) %>">
                 <span
                   aria-label='<%= gettext("Copy Address") %>'
@@ -60,23 +66,20 @@
             </span>
             <!-- bridged token -->
             <%= if @token.bridged do %>
-              <div  style="margin-top: 15px; font-size: 12px;">
-                <span class="mr-4 mb-3 mb-md-0 text-muted" style="margin-right: 30px;">
-                  Token is bridged
-                </span>
+              <div class="view-original-token-container">
                 <%= if Map.has_key?(@token, :foreign_token_contract_address_hash) do %>
                   <a data-test="original_token_contract_address" href=<%= foreign_bridged_token_explorer_link(@token) %> target="_blank">View Original Token <span class="external-token-icon"><%= render BlockScoutWeb.IconsView, "_external_link.html" %></span></a>
                 <% end %>
               </div>
-              <div  style="margin-top: 15px; font-size: 12px;">
-                <%= if Map.has_key?(@token, :custom_metadata) do %>
+              <%= if Map.has_key?(@token, :custom_metadata) do %>
+                <div style="margin-top: 15px; font-size: 12px;">
                   <%= @token.custom_metadata %>
-                <% end %>
-              </div>
+                </div>
+              <% end %>
             <% end %>
           </h1>
 
-          <h3><%= Address.checksum(@token.contract_address_hash) %></h3>
+          <h3 class="token-address-mobile"><%= Address.checksum(@token.contract_address_hash) %></h3>
 
           <div class="d-flex flex-column flex-md-row justify-content-start text-muted">
             <span class="mr-4 mb-3 mb-md-0">
