+++ apps/block_scout_web/priv/gettext/en/LC_MESSAGES/default.po
@@ -405,8 +405,8 @@ msgstr ""
 #, elixir-format
 #: lib/block_scout_web/templates/address/overview.html.eex:34
 #: lib/block_scout_web/templates/address/overview.html.eex:38
-#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:32
-#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:36
+#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:38
+#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:42
 msgid "Copy Address"
 msgstr ""
 
@@ -861,7 +861,7 @@ msgstr ""
 
 #, elixir-format
 #: lib/block_scout_web/templates/tokens/instance/overview/_details.html.eex:10
-#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:26
+#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:32
 msgid "Token Details"
 msgstr ""
 
@@ -1181,8 +1181,8 @@ msgstr ""
 #: lib/block_scout_web/templates/address/overview.html.eex:158
 #: lib/block_scout_web/templates/tokens/instance/overview/_details.html.eex:51
 #: lib/block_scout_web/templates/tokens/instance/overview/_details.html.eex:107
-#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:52
-#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:137
+#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:58
+#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:140
 msgid "QR Code"
 msgstr ""
 
@@ -1291,7 +1291,7 @@ msgstr ""
 #, elixir-format
 #: lib/block_scout_web/templates/address/overview.html.eex:55
 #: lib/block_scout_web/templates/tokens/instance/overview/_details.html.eex:52
-#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:53
+#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:59
 msgid "Show QR Code"
 msgstr ""
 
@@ -1453,7 +1453,7 @@ msgid "Topics"
 msgstr ""
 
 #, elixir-format
-#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:107
+#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:110
 msgid "Total Supply"
 msgstr ""
 
@@ -1610,7 +1610,7 @@ msgstr ""
 #, elixir-format
 #: lib/block_scout_web/templates/tokens/instance/overview/_details.html.eex:16
 #: lib/block_scout_web/templates/tokens/instance/overview/_details.html.eex:20
-#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:87
+#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:90
 msgid "View Contract"
 msgstr ""
 
@@ -1725,8 +1725,8 @@ msgstr ""
 #: lib/block_scout_web/templates/address/overview.html.eex:167
 #: lib/block_scout_web/templates/tokens/instance/overview/_details.html.eex:108
 #: lib/block_scout_web/templates/tokens/instance/overview/_details.html.eex:116
-#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:138
-#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:146
+#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:141
+#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:149
 msgid "Close"
 msgstr ""
 
@@ -1743,7 +1743,7 @@ msgstr ""
 
 #, elixir-format
 #: lib/block_scout_web/templates/tokens/instance/overview/_details.html.eex:76
-#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:95
+#: lib/block_scout_web/templates/tokens/overview/_details.html.eex:98
 msgid "Decimals"
 msgstr ""
 
@@ -2678,7 +2678,7 @@ msgid "Approximate Current Annual Percentage Yield. If you see N/A, please wait 
 msgstr ""
 
 #, elixir-format
-#: lib/block_scout_web/templates/bridged_tokens/index.html.eex:5
+#: lib/block_scout_web/templates/bridged_tokens/index.html.eex:6
 msgid "Bridged Tokens from "
 msgstr ""
